refactor(l1): consolidate tx type tests
