+++ test/tests/common/mod.rs
@@ -2,7 +2,6 @@ mod base64_tests;
 #[cfg(feature = "c-kzg")]
 mod blobs_bundle_tests;
 mod eip7702_authorization_tests;
-mod l1_transaction_types_tests;
 mod logs_bloom_validation_tests;
 mod rkyv_utils_tests;
 mod serde_utils_tests;
